Refactor: organize routes & types - organize Workbook type into types.ts - organize all routes into routing.ts - change navigating logic's url
part of SWM-77
related to #11
