+++ src/app/main/workbook/workbook-cards.tsx
@@ -1,23 +1,20 @@
 'use client';
 
-export type Workbook = {
-  id: string;
-  title: string;
-  description: string;
-  emoji: string;
-  quantity: number;
-  modifiedAt: Date;
-};
+import { Workbook } from '@/app/types';
+import Link from 'next/link';
+import { SUB_ROUTES } from '@/app/routing';
 
 const WorkbookCard = ({ workbook }: { workbook: Workbook }) => {
   return (
-    <div className="w-4/5 p-1 border-black border-[1px]">
-      <div>{workbook.emoji}</div>
-      <div>{workbook.title}</div>
-      <div>{workbook.description}</div>
-      <div>{workbook.quantity}</div>
-      <div suppressHydrationWarning>{workbook.modifiedAt.toString()}</div>
-    </div>
+    <Link href={SUB_ROUTES.workbookDetail(workbook.id)} passHref>
+      <div className="w-4/5 p-1 border-black border-[1px]">
+        <div>{workbook.emoji}</div>
+        <div>{workbook.title}</div>
+        <div>{workbook.description}</div>
+        <div>{workbook.quantity}</div>
+        <div suppressHydrationWarning>{workbook.modifiedAt.toString()}</div>
+      </div>
+    </Link>
   );
 };
