+++ src/app/types.ts
@@ -1,0 +1,8 @@
+export type Workbook = {
+  id: string;
+  title: string;
+  description: string;
+  emoji: string;
+  quantity: number;
+  modifiedAt: Date;
+};
